window: Overhaul UI for GNOME Circle (#14)
A big PR that sizably overhauls the user interface, making Binary more intuitive, robust and easier to read. Thanks to @gregorni for the original design. This PR also comes with converting object names to snake_case for better legibility and consistency with Python and GNOME norms. 

List of relevant changes and commits (see #14 for more):

* window: Flatten blueprint

* main: Credit Gregor Niehl for design overhaul

* window: Move menu button to start of headerbar

* window: Put all elements in main area

* window: Rename input bit counter

* window: Add spacing between elements

* window: Rename entry to input_entry

* window: Replace output label with entry

* window: Set smaller default width

* window: Make dropdown names snake_case

* window: Seperate input and output bit counters 

* window: Replace swap button with arrows

* window: Use snake_case for more signals

* main: Use AdwAboutDialog instead of AboutWindow

* window: Make both input and output immutable

* window: bitsTxt -> bits_text

* window: Make input strings uppercase

This list was trimmed from the full 52 commits, as many were reimplementations of those made on main.